Remove almost all deprecated gravity and anchor-point stuff
Remove some code that is not needed at all. To hold placement of
the previews, instead of gravity, there are two new methods
defined in the Preview class itself: {set,get}_placement() funcs.

Note that, there should be no compatibility issues, as the whole
gravity and anchor-point are deprecated in Clutter 1.12, and the
current version (at the time of current commit) is 1.26.